Model stored on FIrebase and infered by Firebase
Inference is now done by Firebase and not Tensorflow and model is hosted on the Firebase AutoML
