+++ android/app/build.gradle
@@ -1,4 +1,5 @@
 apply plugin: 'com.android.application'
+apply plugin: 'com.google.gms.google-services'
 apply plugin: 'de.undercouch.download'
 
 android {
@@ -57,4 +58,7 @@ dependencies {
     implementation 'androidx.coordinatorlayout:coordinatorlayout:1.0.0'
     implementation 'org.tensorflow:tensorflow-lite:0.0.0-nightly'
     implementation 'androidx.navigation:navigation-runtime:2.2.1'
+    implementation 'com.google.firebase:firebase-analytics:17.2.2'
+    implementation 'com.google.firebase:firebase-auth:19.3.0'
+    implementation 'com.google.firebase:firebase-ml-model-interpreter:22.0.1'
 }
