+++ android/app/src/main/java/com/example/climbingapp/ui/tflite/Classifier.java
@@ -18,7 +18,7 @@ public interface Classifier {
 
     void setUseNNAPI(boolean isChecked);
 
-    public class Recognition {
+    class Recognition {
         private final String id;
         private final String title;
         private final Float confidence;
