+++ android/build.gradle
@@ -5,12 +5,11 @@ buildscript {
         google()
         jcenter()
         mavenLocal()
-
-        
     }
     dependencies {
         classpath 'com.android.tools.build:gradle:3.5.3'
         classpath 'de.undercouch:gradle-download-task:3.4.3'
+        classpath 'com.google.gms:google-services:4.3.3'
         
         // NOTE: Do not place your application dependencies here; they belong
         // in the individual module build.gradle files
@@ -22,7 +21,6 @@ allprojects {
         google()
         jcenter()
         mavenLocal()
-        
     }
 }
